+++ packages/generator/src/GenerateContext/EnhancedInterfaceType.ts
@@ -16,7 +16,7 @@
 
 import type { InterfaceMetadata } from "@osdk/api";
 import type { InterfaceType } from "@osdk/foundry.ontologies";
-import { __UNSTABLE_wireInterfaceTypeV2ToSdkObjectDefinition } from "@osdk/generator-converters";
+import { wireInterfaceTypeV2ToSdkObjectDefinition } from "@osdk/generator-converters";
 import type { EnhanceCommon } from "./EnhanceCommon.js";
 import { EnhancedBase } from "./EnhancedBase.js";
 
@@ -38,7 +38,7 @@ export class EnhancedInterfaceType extends EnhancedBase<InterfaceType> {
   }
 
   getCleanedUpDefinition(v2: boolean): InterfaceMetadata {
-    return __UNSTABLE_wireInterfaceTypeV2ToSdkObjectDefinition(
+    return wireInterfaceTypeV2ToSdkObjectDefinition(
       this.raw,
       v2,
     );