+++ packages/maker/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @osdk/maker
 
+## 0.26.0
+
+### Patch Changes
+
+- Updated dependencies [9eb67e4]
+  - @osdk/api@2.20.0
+  - @osdk/generator-converters.ontologyir@2.20.0
+
 ## 0.25.0
 
 ### Patch Changes