+++ packages/widget.api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @osdk/widget.api
 
+## 3.17.0
+
+### Patch Changes
+
+- Updated dependencies [9eb67e4]
+  - @osdk/api@2.20.0
+
 ## 3.16.0
 
 ### Patch Changes